refactor(core): modularize PlanService, StatusMutationService, and AuthService (Phase 5)
